# raft: Avoid holding unneeded memory in unstable log's entries array