@a-robinson 2017-04-18
cc @bdarnell @petermattis for thoughts on the choice of constant