@bdarnell 2017-04-18
`cap/2` seems like the most natural choice to me, to ensure that we have no more memory in "old" entries than new ones (and it's similar to the common heuristic for arrays to grow by doubling). I think we'd only want to use a higher divisor if there are benchmarks showing the reduced allocations matter (which seems unlikely, since we discard the entire backing array whenever it reaches zero, forcing an allocation on the next append).