@petermattis 2017-04-18
`cap/2` sounds good to me as well without other evidence that a different divisor is better.