@a-robinson 2017-04-18
Sure, we should probably stick with the default of 2 unless we have proof something else is better.